send notifications in addToTeamHelper
